requirements.txt
python-dotenv
aiogram
aiohttp
pillow
